Engine terminating checker should start after exposed
