+++ externals/kyuubi-spark-sql-engine/src/main/scala/org/apache/kyuubi/engine/spark/operation/SparkSQLOperationManager.scala
@@ -50,8 +50,6 @@ class SparkSQLOperationManager private (name: String) extends OperationManager(n
     sessionToSpark.remove(sessionHandle)
   }
 
-  def getOpenSparkSessionCount: Int = sessionToSpark.size()
-
   override def newExecuteStatementOperation(
       session: Session,
       statement: String,
