+++ externals/kyuubi-spark-sql-engine/src/test/scala/org/apache/kyuubi/engine/spark/operation/SparkOperationSuite.scala
@@ -450,14 +450,11 @@ class SparkOperationSuite extends WithSparkSQLEngine with JDBCTests {
     withThriftClient { client =>
       val operationManager = engine.backendService.sessionManager.
         operationManager.asInstanceOf[SparkSQLOperationManager]
-      assert(operationManager.getOpenSparkSessionCount === 0)
 
       val req = new TOpenSessionReq()
       req.setUsername("kentyao")
       req.setPassword("anonymous")
       val tOpenSessionResp = client.OpenSession(req)
-
-      assert(operationManager.getOpenSparkSessionCount === 1)
 
       val tExecuteStatementReq = new TExecuteStatementReq()
       tExecuteStatementReq.setSessionHandle( tOpenSessionResp.getSessionHandle)
@@ -485,7 +482,6 @@ class SparkOperationSuite extends WithSparkSQLEngine with JDBCTests {
       val tCloseSessionResp = client.CloseSession(tCloseSessionReq)
       assert(tCloseSessionResp.getStatus.getStatusCode === TStatusCode.SUCCESS_STATUS)
 
-      assert(operationManager.getOpenSparkSessionCount === 0)
     }
   }
